feat : 글 삭제 기능 추가  S12P21D110-228

See merge request s12-ai-speech-sub1/S12P21D110!87

+++ backend/src/main/java/org/example/backend/community/repository/CommunityRepository.java
@@ -4,15 +4,17 @@ import org.example.backend.community.model.Community;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CommunityRepository extends JpaRepository<Community, Long> {
+public interface CommunityRepository extends JpaRepository<Community, Integer> {
 
     // 카테고리별 조회
     @Query("SELECT c FROM Community c WHERE c.communityCategory = :category")
@@ -28,4 +30,5 @@ public interface CommunityRepository extends JpaRepository<Community, Long> {
 
     @Query("SELECT c FROM Community c WHERE c.articleId = :articleId")
     Optional<Community> findByArticleId(@Param("articleId") Integer articleId);
+
 }
